// src/client/zod.gen.ts
export const zUserCreate = z.object({
    username: z.string().min(3).max(50),
    name: z.optional(z.union([
        z.string().max(100),
        z.null()
    ])),
    profile_photo_url: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    is_active: z.optional(z.boolean()).default(true),
    email: z.email(),
    password: z.optional(z.union([
        z.string().min(8),
        z.null()
    ])),
    google_id: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    google_refresh_token: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    google_access_token: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    google_auth_active: z.optional(z.boolean()).default(false),
    role_name: z.optional(z.union([
        z.enum([
            'admin',
            'manager',
            'user'
        ]),
        z.null()
    ]))
});